AB#23678 move folder location of the bulk approval
