+++ applications/Unity.GrantManager/src/Unity.GrantManager.Web/Pages/BulkApprovals/ApproveApplicationsModal.cshtml
@@ -1,5 +1,5 @@
 ﻿@page
-@model Unity.GrantManager.Web.Pages.GrantApplications.Approvals.ApproveApplicationsModalModel
+@model Unity.GrantManager.Web.Pages.BulkApprovals.ApproveApplicationsModalModel
 @using Microsoft.Extensions.Localization
 @using Unity.GrantManager.Localization
