+++ applications/Unity.GrantManager/src/Unity.GrantManager.Web/Pages/BulkApprovals/ApproveApplicationsModal.css
@@ -17,7 +17,6 @@
 }
 
 
-
 /* This is for the summary modal*/
 
 .table-container {
