+++ applications/Unity.GrantManager/src/Unity.GrantManager.Web/Pages/BulkApprovals/ApproveApplicationsSummaryModal.cshtml
@@ -1,5 +1,5 @@
 ﻿@page
-@model Unity.GrantManager.Web.Pages.GrantApplications.Approvals.ApproveApplicationsSummaryModalModel
+@model Unity.GrantManager.Web.Pages.BulkApprovals.ApproveApplicationsSummaryModalModel
 @using Microsoft.Extensions.Localization
 @using Unity.GrantManager.Localization
 @using Volo.Abp.AspNetCore.Mvc.UI.Bootstrap.TagHelpers.Modal
