+++ applications/Unity.GrantManager/src/Unity.GrantManager.Web/Pages/GrantApplications/Index.cshtml
@@ -24,8 +24,7 @@
             }
         </script>
         <abp-script src="/libs/select2/js/select2.full.min.js" />
-        <abp-script src="/Pages/GrantApplications/Index.js" />
-        <abp-script src="/Pages/GrantApplications/Approvals/ApproveApplicationsModal.js" />
+        <abp-script src="/Pages/GrantApplications/Index.js" />        
     }
     else
     {
@@ -34,8 +33,7 @@
 }
 
 @section styles {
-    <abp-style src="/Pages/GrantApplications/Index.css" />
-    <abp-style src="/Pages/GrantApplications/Approvals/ApproveApplicationsModal.css" />
+    <abp-style src="/Pages/GrantApplications/Index.css" />    
 }
 
 <div class="container-fluid px-2 container-grant-application">
